# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Keystore server",
        "version": "1.0.3",
        "title": "Keystore",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "####@####.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:3000",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/wallet/restore": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Return secret info",
                "description": "Returns secretName, secretVersion \nBut make new secret return /generate",
                "operationId": "restore",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "secretName",
                        "in": "formData",
                        "description": "Input your secretName",
                        "required": true,
                        "type": "string",
                        "default": "dgsafd2415124"
                    },
                    {
                        "name": "mnemonic",
                        "in": "formData",
                        "description": "Input your mnemonic",
                        "required": true,
                        "type": "string",
                        "default": "common tooth lonely custom lab dignitsy bring spike window industry exclude earth"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
        "/keys/create": {
            "post": {
                "tags": [
                    "Keys"
                ],
                "summary": "Return key info",
                "description": "Returns keyInfo, crv, Buffer X, Buffer Y",
                "operationId": "create",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "keyName",
                        "in": "formData",
                        "description": "Create keyName",
                        "required": true,
                        "type": "string",
                        "default": "testkeygg"
                    },
                    {
                        "name": "keyType",
                        "in": "formData",
                        "description": "Input keyType",
                        "required": true,
                        "type": "string",
                        "default": "EC"
                    }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/keys/import": {
            "put": {
                "tags": [
                    "Keys"
                ],
                "summary": "Return key info",
                "description": "Returns keyInfo, crv, Buffer X, Buffer Y",
                "operationId": "import",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "keyName",
                        "in": "formData",
                        "description": "Import keyName",
                        "required": true,
                        "type": "string",
                        "default": "testkeygg"
                    },
                    {
                        "name": "d",
                        "in": "formData",
                        "description": "Input privateKey",
                        "required": true,
                        "type": "string",
                        "default": "Vgyk4k6SQnGJ2rMtR19XTBYBik3pfF9+aJXAfjk7hoY="
                    },
                    {
                        "name": "x",
                        "in": "formData",
                        "description": "Input publicX",
                        "required": true,
                        "type": "string",
                        "default": "zXcMOqKdEB6a/AR7ZjI6j15vsQXB46XV4nVye4Zh4PY="
                    },
                    {
                        "name": "y",
                        "in": "formData",
                        "description": "Input publicY",
                        "required": true,
                        "type": "string",
                        "default": "c6UhjMtcT0zt2HIxpBgHj/Ns4JJ/TNkgXJAuypdZrdI="
                    }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/keys/get": {
            "get": {
                "tags": [
                    "Keys"
                ],
                "summary": "Return key info",
                "description": "Returns keyInfo, crv, Buffer X, Buffer Y",
                "operationId": "get",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "keyName",
                        "in": "query",
                        "description": "Input your keyName",
                        "required": true,
                        "type": "string",
                        "default": "testkey"
                    },
                    {
                        "name": "keyVersion",
                        "in": "query",
                        "description": "Input your keyVersion",
                        "required": true,
                        "type": "string",
                        "default": "1d59b68fe3184d499de2c52524c3a833"
                    }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/keys/getS": {
            "get": {
                "tags": [
                    "Keys"
                ],
                "summary": "Return secretInfo",
                "description": "Return secretInfo",
                "operationId": "getS",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "secretName",
                        "in": "query",
                        "description": "Input your secretName",
                        "required": true,
                        "type": "string",
                        "default": "dgsafd2415"
                    },
                    {
                        "name": "secretVersion",
                        "in": "query",
                        "description": "Input your secretVersion",
                        "required": true,
                        "type": "string",
                        "default": "b111414890e4402e8ffb784096959fe3"
                    }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/keys/delete": {
            "delete": {
                "tags": [
                    "Keys"
                ],
                "summary": "Return key info",
                "description": "Returns keyInfo, crv, Buffer X, Buffer Y",
                "operationId": "delete",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "keyName",
                        "in": "formData",
                        "description": "Input your keyName",
                        "required": true,
                        "type": "string",
                        "default": "testkeygg"
                    }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/keys/deleteS": {
            "delete": {
                "tags": [
                    "Keys"
                ],
                "summary": "Return secretInfo",
                "description": "Return secretInfo",
                "operationId": "deleteSecret",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "secretName",
                        "in": "formData",
                        "description": "Input your secretName",
                        "required": true,
                        "type": "string",
                        "default": "dgsafd"
                    }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/keys/sign": {
            "post": {
                "tags": [
                    "Keys"
                ],
                "summary": "Return signature",
                "description": "Return signature",
                "operationId": "sign",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "keyName",
                        "in": "formData",
                        "description": "Input your keyName",
                        "required": true,
                        "type": "string",
                        "default": "testkey"
                    },
                    {
                        "name": "keyVersion",
                        "in": "formData",
                        "description": "Input your keyVersion",
                        "required": true,
                        "type": "string",
                        "default": "1d59b68fe3184d499de2c52524c3a833"
                    },
                    {
                        "name": "value",
                        "in": "formData",
                        "description": "Input your privateKey",
                        "required": true,
                        "type": "string",
                        "default": "Vgyk4k6SQnGJ2rMtR19XTBYBik3pfF9+aJXAfjk7hoY="
                    },
                    {
                        "name": "algorithm",
                        "in": "formData",
                        "description": "Input algorithm",
                        "required": true,
                        "type": "string",
                        "default": "ECDSA256"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/keys/verify": {
            "post": {
                "tags": [
                    "Keys"
                ],
                "summary": "Return value",
                "description": "Return value",
                "operationId": "verify",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "keyName",
                        "in": "formData",
                        "description": "Input your keyName",
                        "required": true,
                        "type": "string",
                        "default": "testkey"
                    },
                    {
                        "name": "keyVersion",
                        "in": "formData",
                        "description": "Input your keyVersion",
                        "required": true,
                        "type": "string",
                        "default": "1d59b68fe3184d499de2c52524c3a833"
                    },
                    {
                        "name": "digest",
                        "in": "formData",
                        "description": "Input your privateKey",
                        "required": true,
                        "type": "string",
                        "default": "Vgyk4k6SQnGJ2rMtR19XTBYBik3pfF9+aJXAfjk7hoY="
                    },
                    {
                        "name": "algorithm",
                        "in": "formData",
                        "description": "Input algorithm",
                        "required": true,
                        "type": "string",
                        "default": "ECDSA256"
                    },
                    {
                        "name": "signature",
                        "in": "formData",
                        "description": "Input signature",
                        "required": true,
                        "type": "string",
                        "default": "c41cad91ebaa252e346b63275421b418299485a49cfd27a1b4b85aae22a1eeeb9806daf859918382873f60cf4eb2dd24cdaf5d680cb1638e4fa87f47fb287187"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/keys/backup": {
            "post": {
                "tags": [
                    "Keys"
                ],
                "summary": "Return keyBundle",
                "description": "Return keyBundle",
                "operationId": "backup",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "keyName",
                        "in": "formData",
                        "description": "Input your keyName",
                        "required": true,
                        "type": "string",
                        "default": "testkeygg"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    }
}